added toggle switches

+++ BouncingBall/src/UI/RuleChangeEventArgs.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace BouncingBall.UI;
 
-public class RuleChangeEventArgs(RuleType rule, float value) : EventArgs {
+public class RuleChangeEventArgs(RuleType rule, int value) : EventArgs {
     
     public readonly RuleType Rule = rule;
